Refactor common state in intake reducers
